# swagger.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Local server"
    }
  ],
  "info": {
    "version": "1.0.0",
    "title": "MTO Backend Project"
  },
...
  "paths": {
...
    "/admin/merchants/{id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "description": "Get the specific merchant with the given id",
        "operationId": "getMerchantById",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Merchant"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "id must be a string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No data found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoDataFoundMessage"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/agents": {
      "get": {
        "tags": [
          "admin"
        ],
        "description": "Get all agents which are available on the marketplace",
        "operationId": "getAgents",
        "parameters": [
          {
            "$ref": "#/components/parameters/page"
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/walletAddressToFilter"
          },
          {
            "$ref": "#/components/parameters/agentStatus"
          },
          {
            "$ref": "#/components/parameters/fromScore"
          },
          {
            "$ref": "#/components/parameters/toScore"
          },
          {
            "$ref": "#/components/parameters/sortBy"
          },
          {
            "$ref": "#/components/parameters/order"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Agent"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidFilterMessage"
                }
              }
            }
          },
          "404": {
            "description": "No data found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoDataFoundMessage"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "id": {
        "in": "path",
        "name": "id",
        "schema": {
          "type": "integer"
        },
        "required": true
      },
...
      "walletAddressToFilter": {
        "in": "query",
        "name": "walletAddress",
        "schema": {
          "type": "string",
          "example": "0x12345abcde7890"
        }
      },
      "agentStatus": {
        "in": "query",
        "name": "status",
        "schema": {
          "type": "string",
          "enum": [
            "init",
            "waiting",
            "review",
            "pending_01",
            "pending_02",
            "earned",
            "lost",
            "ban"
          ]
        }
      },
      "page": {
        "in": "query",
        "name": "page",
        "schema": {
          "type": "number"
        }
      },
      "limit": {
        "in": "query",
        "name": "limit",
        "schema": {
          "type": "number"
        }
      },
...
      "fromScore": {
        "in": "query",
        "name": "fromScore",
        "schema": {
          "type": "number"
        }
      },
      "toScore": {
        "in": "query",
        "name": "toScore",
        "schema": {
          "type": "number"
        }
      },
...
      "sortBy": {
        "in": "query",
        "name": "sortBy",
        "schema": {
          "type": "string"
        }
      },
      "order": {
        "in": "query",
        "name": "order",
        "schema": {
          "type": "string",
          "enum": [
            "DESC",
            "ASC"
          ]
        }
      }
    },
    "schemas": {
      "Agent": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "dispute": {
            "type": "integer",
            "allOf": [
              {
                "$ref": "#/components/schemas/Dispute"
              }
            ]
          },
          "walletAddress": {
            "type": "string",
            "example": "0x12345abcde7890"
          },
          "score": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "init",
              "waiting",
              "review",
              "pending_01",
              "pending_02",
              "earned",
              "lost",
              "ban"
            ]
          }
        }
      },
...
      "Dispute": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "description": {
            "type": "string"
          },
          "approvedCount": {
            "type": "integer"
          },
          "disapprovedCount": {
            "type": "integer"
          },
          "reviewCount": {
            "type": "integer"
          },
          "criteriaCount": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "default": "init",
            "enum": [
              "init",
              "waiting",
              "review",
              "fail",
              "win"
            ]
          }
        }
      },
      "Merchant": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "walletAddress": {
            "type": "string",
            "example": "0x12345abcde7890"
          },
          "shippingAddress": {
            "type": "string"
          },
          "externalLink": {
            "type": "string"
          }
        }
      },
...
      "NoDataFoundMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "No data found"
          }
        }
      },
      "InvalidFilterMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "page must be greater than or equal to 1"
          }
        }
      }
    },
...
  }
}